Chore: Move Gatsby Node back to root
Gatsby should clear cache when you make changes to any of these files.
But it doesn't detect these changes when not in the default location.
